+++ packages/datadog_flutter_plugin/lib/datadog_flutter_plugin.dart
@@ -74,10 +74,9 @@ class DatadogSdk {
 
   final Map<Type, DatadogPlugin> _plugins = {};
 
-  /// A list of first party hosts for tracing. Note that this is an unmodifiable
-  /// list. If you need to add a host, call the setter for [firstPartyHosts]
+  /// An unmodifiable list of first party hosts for tracing.
   List<String> get firstPartyHosts => List.unmodifiable(_firstPartyHosts);
-  set firstPartyHosts(List<String> value) {
+  void _setFirstPartyHosts(List<String> value) {
     _firstPartyHosts = value;
     if (value.isNotEmpty) {
       // pattern = "^(.*\\.)*tracedHost1$|tracedHost2$|...$"
@@ -150,7 +149,7 @@ class DatadogSdk {
     configuration.additionalConfig[DatadogConfigKey.source] = 'flutter';
     configuration.additionalConfig[DatadogConfigKey.version] = version;
 
-    firstPartyHosts = configuration.firstPartyHosts;
+    _setFirstPartyHosts(configuration.firstPartyHosts);
 
     await _platform.initialize(configuration, logCallback: _platformLog);