+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Send `firstPartyHosts` to Native SDKs during initialization. Make
+  `firstPartyHosts` property on read only `DatadogSdk` read only. 
+
 ## 1.0.0-alpha.2
 
 * Cancel spans on DatadogTrackingHttpClient when RUM is enabled (prevent spans